funcion/11274-Sincronizar-repo-git-con-SVN

See merge request ulises/sirtap!2

+++ ulises-ta/Tools/pjproject/pjsip/src/test/test.c
@@ -379,8 +379,8 @@ on_return:
     PJ_LOG(3,(THIS_FILE, ""));
  
     pj_thread_get_stack_info(pj_thread_this(), &filename, &line);
-    PJ_LOG(3,(THIS_FILE, "Stack max usage: %u, deepest: %s:%u", 
-	              pj_thread_get_stack_max_usage(pj_thread_this()),
+    PJ_LOG(3,(THIS_FILE, "Stack max usage: %llu, deepest: %s:%u", 
+	              (unsigned long long) pj_thread_get_stack_max_usage(pj_thread_this()),
 		      filename, line));
     if (rc == 0)
 	PJ_LOG(3,(THIS_FILE, "Looks like everything is okay!.."));
